fix: custom roles and groups definition with sea-orm
